Remove `IAsyncInitialize` interface.
Resolve #12913
