+++ framework/src/Volo.Abp.AspNetCore.Components.Web/Volo/Abp/AspNetCore/Components/Web/Security/AbpComponentsClaimsCache.cs
@@ -8,10 +8,9 @@ using Volo.Abp.DependencyInjection;
 namespace Volo.Abp.AspNetCore.Components.Web.Security;
 
 [ExposeServices(
-    typeof(AbpComponentsClaimsCache),
-    typeof(IAsyncInitialize)
+    typeof(AbpComponentsClaimsCache)
 )]
-public class AbpComponentsClaimsCache : IScopedDependency, IAsyncInitialize
+public class AbpComponentsClaimsCache : IScopedDependency
 {
     public ClaimsPrincipal Principal { get; private set; }
