+++ framework/src/Volo.Abp.AspNetCore.Mvc.Client.Common/Volo/Abp/AspNetCore/Mvc/Client/ICachedApplicationConfigurationClient.cs
@@ -3,7 +3,7 @@ using Volo.Abp.AspNetCore.Mvc.ApplicationConfigurations;
 
 namespace Volo.Abp.AspNetCore.Mvc.Client;
 
-public interface ICachedApplicationConfigurationClient : IAsyncInitialize
+public interface ICachedApplicationConfigurationClient
 {
     Task<ApplicationConfigurationDto> GetAsync();
