+++ framework/src/Volo.Abp.AspNetCore.Mvc.Client/Volo/Abp/AspNetCore/Mvc/Client/MvcCachedApplicationConfigurationClient.cs
@@ -13,9 +13,8 @@ namespace Volo.Abp.AspNetCore.Mvc.Client;
 
 [ExposeServices(
     typeof(MvcCachedApplicationConfigurationClient),
-    typeof(ICachedApplicationConfigurationClient),
-    typeof(IAsyncInitialize)
-    )]
+    typeof(ICachedApplicationConfigurationClient)
+)]
 public class MvcCachedApplicationConfigurationClient : ICachedApplicationConfigurationClient, ITransientDependency
 {
     protected IHttpContextAccessor HttpContextAccessor { get; }
